Add Devise gem
This gem allows us to facilitate the sessions of a user within the
project, providing the necessary resources for do it.
This adds the necessary congurations so that Devise works
correctly in the project
Finally this creates the views for the sessions that Devise handles,
these are generated to create a user, edit, initiate session
among other configurations.

Co-authored-by: JPablo90 <jp.garciamurillo@gmail.com>