Create relation of models
This creates the relations of models, in such a way that the user
can have many companies but a company only belongs to one user and
a company can have an address but an address only belongs to one
company.
The migrations are modified to add the relations in the tables of the
database, also the database is dropped to create it again to add the
new fields to the schema by a new migration.

Co-authored-by: JPablo90 <jp.garciamurillo@gmail.com>