Modify navbar position and background
Adds a background with the text 'CANACO', this in order to set the navbar
at middle of the background, this is a better version with clearer letters
and different font.
Also this adds a button for sign in or show your name in case that the user
is already logged in.
It changes the test in order to add another one for the new interaction.
Adds the new navbar to the views that used an old version
of the same.
Also removes the old styles for that views because before
the navbar and the background were differents elements.

Co-authored-by: Orlando Sotelo <48534682+LandoSOt@users.noreply.github.com>