Delete unnecessary elements
Delete images from assets, this is because the images are not
required or there are better versions of these.
Also deletes home helper and empty test for this helper, in order to keep
the project without unnecessary files and also the helper is no longer
needed since a long time ago.

Co-authored-by: Orlando Sotelo <48534682+LandoSot@users.noreply.github.com>