Add confirmable email to users
This adds settings in the user model to confirm a user
account by email. These configurations are necessary
to ensure that the user's information is correct.

A new view is created and customized to inform the user
that an email will be sent to confirm their data.
Is modified the route to which we are redirected by
default after correct signup, this is necessary in
order to access this view.
The 'letter_opener' gem is added to simulate the
sending of an email locally.

Co-authored-by: OswaldoPineda <carlos-1502010@hotmail.com>