Fix errors in the homepage banners
This corrects errors in the design view of the banners on the
home page to follow the design provided.
In addition, the name of some tags in the HTML file is modified
to give a better description of what they do.
Finally, some HTML files are renamed by a suggestion of the
linters tests to follow Ruby on Rails standards.