Create subcategories view
The subcategories controller is created to generate
an action when the user selects a category and render
to subcategories template.
This creates the path to the view of subcategories to
generate the link to it.
The html buy view is modified to update the path when
the user clicks on a category.
Also, is performed the customize of this view, it adds
scss styles and some properties of boostrap, in order
to follow the design provided
Finally, this creates the unit test to ensure that the
subcategory template is rendered.