Improve error alerts in forms
This generates an initializer to add the 'is-invalid'
class at the fields that were filling incorrectly
after of submit.
Two views in the project are modified to add the error
messages under the fields were filling incorrectly,
the messages are added through a helper.
Also, the 'sign-up.js' file is modified to get the
correct value of the neighborhood in the address model,
to avoid that this field was erased if the user is wrong.
Finally, the 'application.css' is fixed because an ';'
is missing.

Co-authored-by: OswaldoPineda <carlos-1502010@hotmail.com>