Change some hard codes by labels.
In some views, the hard code is changed by the labels
defined in es.yml file to avoid modifying many files
when the meaning of some text changes.
In addition, an scss file is added to put an asterisk
after the text of the labels in the fields that are
important to fill in the registration view.