Complete general front-end
This customizes the views in the app to follow the
design provide. Two new scss files are added to generate
the colors that will be used throughout the project,
these colors are saved in scss variables to facilitate
the personalization of the views, while some scss files
are modified or deleted because they are no longer
necessary.
On the other hand, are added the icons and the images to
allow us to stick to the aforementioned design.
Also, some views are modified to remove the footer and
the navigation bar.
Finally, are modified the text in some views by a request
of the designer, and the text in some spec files are changed
to correct the errors that arose when modifying some files.