Configurate the middleware in rails
Is necesary to modify the routes to avoid have
Rails’s default error pages, since these are boring
and bland.
The hook of Rails is modified to creating a custom
error pages (404, 422 and 500).
Also, a new controller and its unit test are created.
This controller is in charge to redirect to the new
views and follow the configuration to custom the
error pages.
Finally the new error views are created and the default
views are removed, the application view is modified to
remove the header and navigation bar from error views.