Create the general view from the chat
Include the necessary icons for the raiting of the sellers.
Include the scss file import into aplication.scss.
Create the scss styles into chat.scss file.
Generate the controller for the view.
Create partials to keep and order in the chat view.
Generate the index for the view.
Configure the route for the view into the routes file.
Include the test for the controller and the view files.