Customize email templates with scss styles
This adds the premailer gem and its dependencies to keep
the css styles in the email clients, since some of these have
bad support for stylesheets.

The default email dispatcher is deleted because it is no longer
necessary. A new dispatcher is added, the wich extends of
devise mailer  to create our own personalized email templates.

The css file is created and required in the mailer.html. Finally,
this file is configured as a precompile in the 'assets' to apply
the styles before sending the emails to the clients. Also, the
information in this view is translated into Spanish and such
information is modified to give users personalized messages.