Create search by category and subcategory
This creates the controllers to search petitions by category and
subcategory also creates its routes and add the links in the categories dropdown.

If the search is through these controllers creates a variable to save
the title and could print it on the view, also modify the title in the search
view to print the written text or the category or subcategory title if exist.