Converts the CANACO logo to button
The CANACO logo is becomes to button in those views that do not
have footer and navbar, to allow the user to return to the
homepage.
This logo is added in the registry view since it did not exist.
Also, some modifications in the scss styles are necessaries to
maintain the given design.