Edit the necessary files and replace the old icons for the new style
Replace the old icons for the new ones.
Edit the scss files to be responsive with smaller screens.
Edit the html view file to fix some errors with the responsive
on small screens.
Include color variables for the scss files.
Include an error in the I18n file for missing images.