Add the require configuration for load images
Include into users model file, the variables, credentials and routes
required to save the avatar picture of the users.
Add the required variables into development and production rb files.
Include the avatar key allowed key in application_controller file.
Edit the navbar and registration edit views to include the profile picture default
or the picture uploaded for the current user.
Edit a test according to the edit view changes.
Create a helper for avatar options.

Co-authored-by: Oswaldo Pineda <carlos-1502010@hotmail.com>