Generate mobile search view
This creates a new view for searching by categories and
subcategories on mobile devices, in order to have a way to access
the search view. This has the style sheet file and add this to the
application file. Finally, unit and integration tests are performed
to check the correct functioning of the new view and its route.