Load Dropzone just on sales
Load dropzone plugin just in sales view.

By default the project always includes the js Dropzone plugin,
and the web explorer was always calling it,
this always raises an error from element not found in views not needed.

To solve this, just add an if with the id from dropzone,
this avoids the automatic load and just loads when the element is needed.