Change the Chat View
Change all the conversations views to match the design on zeplin.
Add color variables to the _variables.scss file.
Add the attachment icon to the chat.
Change the responsive design to look better.
Add validations for the top bar when there is no selected chat.
Change some of the texts displayed.
Add the new star icons designs.